+++ hdforce/GetTestsType.py
@@ -4,7 +4,7 @@ import os
 import datetime
 import pandas as pd
 # Package imports
-from .utils import responseHandler, logger, ConfigManager, deprecated
+from .utils import responseHandler, logger, ConfigManager, deprecated, dtConverter
 from .AuthManager import AuthManager
 # Enable deprecation warnings globally
 import warnings
@@ -101,17 +101,24 @@ def GetTestsType(typeId: str, from_: int = None, to_: int = None, sync: bool = F
     # Create blank Query list to handle parameters
     query = {}
 
+    # Convert from_ and to_ to epoch timestamps
+    if from_ is not None:
+        from_ = dtConverter(from_)
+    if to_ is not None:
+        to_ = dtConverter(to_)
+
     # Evaluate from and to dates for Sync command
     if sync is True:
         if from_ is not None:
             query['syncFrom'] = from_
         if to_ is not None:
             query['syncTo'] = to_
-    elif sync is False:
+    else:
         if from_ is not None:
             query['from'] = from_
         if to_ is not None:
             query['to'] = to_
+
 
     # Check typeId
     type_ids = {
@@ -183,6 +190,11 @@ def GetTestsType(typeId: str, from_: int = None, to_: int = None, sync: bool = F
         df.attrs['Last Sync'] = int(data['lastSyncTime'])
         df.attrs['Last Test Time'] = int(data['lastTestTime'])
         df.attrs['Count'] = int(data['count'])
+
+        # Add Last Sync column
+        last_sync_time = int(data['lastSyncTime'])
+        df['last_sync_time'] = last_sync_time
+        
         logger.info(f"Request successful. Returned {df.attrs['Count']} {df.attrs['Test Type Name']} tests.")
         return df