+++ hdforce/GetTypes.py
@@ -1,18 +1,11 @@
 # Dependencies -----
-import requests
-import os
-import datetime
 import pandas as pd
-# Package imports
-from .utils import logger, ConfigManager
-from .AuthManager import AuthManager
 
 # -------------------- #
 # Get Test Types
 
-
 def GetTypes() -> pd.DataFrame:
-    """Get the test type names and IDs from an API.
+    """Get the test type names and IDs.
 
     Returns
     -------
@@ -20,84 +13,21 @@ def GetTypes() -> pd.DataFrame:
         A Pandas DataFrame containing the test types, with columns:
         - id: The unique identifier for each test type.
         - name: The name of each test type.
-
-    Raises
-    ------
-    Exception
-        If the HTTP response status is not 200, indicating an unsuccessful API request,
-        or if there is a failure in parsing the JSON response.
+        - abbreviation: The abbreviation for each test type.
     """
-    # Retrieve Access Token and check expiration
-    a_token = ConfigManager.get_env_variable("ACCESS_TOKEN")
-    tokenExp = int(ConfigManager.get_env_variable("TOKEN_EXPIRATION"))
+    test_data = {
+        "canonicalId": [
+            "7nNduHeM5zETPjHxvm7s", "QEG7m7DhYsD6BrcQ8pic", "2uS5XD5kXmWgIZ5HhQ3A",
+            "gyBETpRXpdr63Ab2E0V8", "5pRSUQVSJVnxijpPMck3", "pqgf2TPUOQOQs6r0HQWb",
+            "r4fhrkPdYlLxYQxEeM78", "ubeWMPN1lJFbuQbAM97s", "rKgI4y3ItTAzUekTUpvR",
+            "4KlQgKmBxbOY6uKTLDFL", "umnEZPgi6zaxuw0KhUpM"
+        ],
+        "name": [
+            "Countermovement Jump", "Squat Jump", "Isometric Test", "Drop Jump",
+            "Free Run", "CMJ Rebound", "Multi Rebound", "Weigh In", "Drop Landing",
+            "TS Free Run", "TS Isometric Test"
+        ],
+        "abbreviation": ["CMJ", "SJ", "ISO", "DJ", "FR", "CMJR", "MR", "WI", "DL", "TSFR", "TSISO"]
+    }
 
-    # get current time in timestamp
-    now = datetime.datetime.now()
-    nowtime = datetime.datetime.timestamp(now)
-
-    # Validate refresh token and expiration
-    if a_token is None:
-        logger.error("No Access Token found.")
-        raise Exception("No Access Token found.")
-    elif int(nowtime) >= tokenExp:
-        logger.debug(f"Token Expired: {datetime.datetime.fromtimestamp(tokenExp)}")
-        # authenticate
-        try:
-            AuthManager(
-                region=ConfigManager.region,
-                authMethod=ConfigManager.env_method,
-                refreshToken_name=ConfigManager.token_name,
-                refreshToken=ConfigManager.refresh_token,
-                env_file_name=ConfigManager.file_name
-            )
-            # Retrieve Access Token and check expiration
-            a_token = ConfigManager.get_env_variable("ACCESS_TOKEN")
-            logger.debug("New ACCESS_TOKEN retrieved")
-            tokenExp = int(ConfigManager.get_env_variable("TOKEN_EXPIRATION"))
-            logger.debug("TOKEN_EXPIRATION retrieved")
-            if a_token is None:
-                logger.error("No Access Token found.")
-                raise Exception("No Access Token found.")
-            elif int(nowtime) >= tokenExp:
-                logger.debug(f"Token Expired: {datetime.datetime.fromtimestamp(tokenExp)}")
-                raise Exception("Token expired")
-            else:
-                logger.debug(f"New Access Token valid through: {datetime.datetime.fromtimestamp(tokenExp)}")
-                pass
-        except ValueError:
-            logger.error("Failed to authenticate. Try AuthManager")
-            raise Exception("Failed to authenticate. Try AuthManage")
-    else:
-        logger.debug(f"Access Token retrieved. expires {datetime.datetime.fromtimestamp(tokenExp)}")
-
-    # API Cloud URL
-    url_cloud = os.getenv("CLOUD_URL")
-
-    # Construct the API endpoint URL
-    url = f"{url_cloud}/test_types"
-
-    # Setup the authorization headers for the API request
-    headers = {"Authorization": f"Bearer {a_token}"}
-
-    # Send the GET request to the API
-    logger.debug("GET request sent for Test Types.")
-    response = requests.get(url, headers=headers)
-
-    # Check if the API response was successful
-    if response.status_code != 200:
-        logger.error(f"Error {response.status_code}: {response.reason}")
-        raise Exception(f"Error {response.status_code}: {response.reason}")
-
-    # If successful
-    try:
-        data = response.json()
-        df = pd.DataFrame.from_records(data)
-
-        df.attrs['Count'] = int(len(df.index))
-        count = str(len(df.index))
-        logger.info(f"Request successful. Types returned: {count}")
-        return df
-
-    except ValueError:
-        logger.error("Failed to parse JSON response or no data returned.")
-        raise Exception("Failed to parse JSON response or no data returned.")
+    return pd.DataFrame(test_data)