+++ hdforce/__init__.py
@@ -19,3 +19,8 @@ from .GetAthletes import GetAthletes
 from .GetGroups import GetGroups
 from .GetTeams import GetTeams
 from .GetTags import GetTags
+from .BuildDB import BuildDB
+from .SyncDB import SyncDB
+
+# Load Metric Dictionary
+from .utils import Metrics